Added control group for node font hyper-link.
